Fix c++ syntax error/typo & Update CMakeLists.txt
change in Classes/external/json.h is workaround to
 corresponding gcc regression bug.

+++ Classes/NonGameplayScenes/LocationSelectScene.cpp
@@ -165,7 +165,7 @@ LocationSelectScene::onEnter()
             rows.push_back(Value(_columnsInRow));
             _size -= _columnsInRow;
         } else {
-            rows.push_back(Value(_size));
+            rows.push_back(Value((double)_size));
             _size = 0;
         }
     }
